Refactor 'normalized_args' to common 'utils.rs'
